assignment_2/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contextlib2==21.6.0
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-ninja==1.3.0
django-ninja-extra==0.21.4
django-ninja-jwt==5.3.3
django-redis==5.4.0
gunicorn==23.0.0
injector==0.22.0
kombu==5.4.0
packaging==24.1
prompt_toolkit==3.0.47
psycopg==3.2.1
psycopg-binary==3.2.1
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
